update model + schema format
