--- server/schemas/network.js
@@ -1,11 +1,0 @@
-const mongoose = require('mongoose')
-
-let networkSchema = mongoose.Schema({
-  totalTokens: Number,
-  totalReputation: Number,
-  currentPrice: Number,
-  ethPrice: Number,
-  weiBal: Number
-})
-
-module.exports = networkSchema
